NSString+Encrypt NSData+Base64

+++ categories/Foundation/NSString/NSString+Hash.m
@@ -48,7 +48,7 @@
     NSData *messageData = [self dataUsingEncoding:NSUTF8StringEncoding];
     NSMutableData *mutableData = [NSMutableData dataWithLength:CC_SHA1_DIGEST_LENGTH];
     CCHmac(kCCHmacAlgSHA1, keyData.bytes, keyData.length, messageData.bytes, messageData.length, mutableData.mutableBytes);
-    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:mutableData.length];
+    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:(int)mutableData.length];
 }
 - (NSString *)hmacSHA256StringWithKey:(NSString *)key
 {
@@ -56,7 +56,7 @@
     NSData *messageData = [self dataUsingEncoding:NSUTF8StringEncoding];
     NSMutableData *mutableData = [NSMutableData dataWithLength:CC_SHA256_DIGEST_LENGTH];
     CCHmac(kCCHmacAlgSHA256, keyData.bytes, keyData.length, messageData.bytes, messageData.length, mutableData.mutableBytes);
-    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:mutableData.length];
+    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:(int)mutableData.length];
 }
 - (NSString *)hmacSHA512StringWithKey:(NSString *)key
 {
@@ -64,7 +64,7 @@
     NSData *messageData = [self dataUsingEncoding:NSUTF8StringEncoding];
     NSMutableData *mutableData = [NSMutableData dataWithLength:CC_SHA512_DIGEST_LENGTH];
     CCHmac(kCCHmacAlgSHA512, keyData.bytes, keyData.length, messageData.bytes, messageData.length, mutableData.mutableBytes);
-    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:mutableData.length];
+    return [self stringFromBytes:(unsigned char *)mutableData.bytes length:(int)mutableData.length];
 }
 #pragma mark - Helpers
 - (NSString *)stringFromBytes:(unsigned char *)bytes length:(int)length
